+++ packages/editor-ui/src/composables/useAutocompleteTelemetry.ts
